[update] main.pyからdetail_windowを開ける＋Widgetクラスを分離
ボタンを押したときに
self._detail_window = dw.DetailWindow(self, widget.get_subject(), self._dw_is_opened)
self._detail_window.show_window()
でdetail_windowを開けるようにした。
また、自分でdetail_window.pyのコンストラクタ等と変更して実行したところ、detail_windowが閉じれなかった。原因は不明（たぶんdetail_window.py側の問題？）
rootを使えば親ウィジェットの関数が使えることを発見したので、Widgetクラスをwidget.pyに移動させた。
